requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
starlette==0.38.5

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.3

# AI/ML Models - STT
openai-whisper==20231117
torch>=2.1.0
torchaudio>=2.1.0

# AI/ML Models - RAG
transformers>=4.42.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3

# ElevenLabs TTS
elevenlabs>=1.0.0

# Audio Processing
soundfile>=0.12.1
librosa>=0.10.1
pydub>=0.25.1
webrtcvad>=2.0.10

# Additional utilities
scipy>=1.11.4

